2016-2017 Constitutional Amendment Meeting
RCT Meeting Notes
Purpose: Constitutional Amendment Meeting
Date: July 9, 2017
Location: KHE125
Attendees: Janine Fernandes, Bryan Chu, Mieka Shade, Bronwen Spolsky

Added content is bounded by ^ symbols, removed content is bounded by ++
symbols, replacement content is in the form ^new text^:[++old text++]
Decisions:
Constitutional Amendment Meeting Results

1) APPROVED. Aye: 4, Nay: 0

^Article I^:[++ARTICLE I++]  Name
^Article II^:[++ ARTICLE II++]  Mission Statement
^Article III^:[++ ARTICLE III++]  Artistic Vision

2) APPROVED. Aye: 3, Nay: 0, Abstain: 1

Article VI, Section I, Subsection 1
Executive-At-Large  An applicant is only eligible for consideration if
they are a formal member of RCT. Executives-at-Large are executives
operating on the Executive Board and within the power structure of RCT
without formal titles or responsibilities. However, Executives-at-Large
are the hands, eyes, ears, and sometimes legs of the group. The
executives-at-large work closely with the General Manager to realize
^actions^:[++events++], though of course they may work with the Artistic
Director and Treasurer as well. Executives-at-Large are encouraged and
expected to either come up with or help write at least one
^action^:[++event++] proposal in coordination with the Artistic
Director, partly to help train new executives for leadership in
subsequent years, but mostly to enrich the community as a whole with
varied events. Executives-at-Large will attend meetings, help procure
properties and resources, and provide input. Executives-at-Large will
not be expected to provide input on or make major decisions, but their
input is always appreciated.

Article VI, Section I, Subsection 2, i
Artistic Director - The Artistic Director is responsible for the
guidance and direction of the group. The Artistic Director must have a
clear vision on where s/he would like to take this group. In more
practical terms, the Artistic Director is responsible for overall
planning. The Artistic Director will write detailed ^action^:[++event++]
proposals, covering every aspect of the ^action^:[++event++], and will
make final decisions on matters. During play submission review, the
Artistic Director will lead the decision-making process. While the
Artistic Director has final say in what should happen in the group, this
only matters when there is no consensus. The Artistic Director should
not need to be making final decisions if there is already a majority.

Article VI, Section I, Subsection 2, ii
General Manager- The General Manager is responsible for executing and
procuring the requirements detailed in the Artistic Director's
^action^:[++event++] proposals. The General Manager will set deadlines,
do check-ups, and utilize the executives-at-large (and other resources
of RCT) to accomplish the goals set by the Artistic Director. The
General Manager must ensure that deadlines are being met. The General
Manager may act as Artistic Director (and vice-versa) if the Artistic
Director is absent or unavailable.

Article VII, Section I
Section I  The Treasurer, General Manager, and Artistic Director must be
in agreement for an ^action^:[++event++] proposal to be eligible to be
approved for action.

Article VII, Section II
Section II  An event proposal may only be approved for action after
receiving a majority vote at an ^action^:[++event++] proposal meeting. A
meeting cannot be held within 72 hours of the decision to hold it.

3) APPROVED. Aye: 3, Nay: 0, Abstain: 1

^Article IV, Section I^
^Section I - An action is defined as any formal work, activity, or
operation planned and/or undertaken by the RCT Executive Team. Actions
may include, but are not limited to, events (such as pub nights,
fundraisers, and/or the end-of-year show), organizational projects, or
administrative tasks, and are defined as such at the discretion of the
current season's executive team.
i. Any action defined as such must have an accompanying action proposal,
regardless of its approval status (see Article VII).
ii. An action proposal specifically refers to a written plan and/or
outline describing an action.^

4) APPROVED. Aye: 3, Nay: 0, Abstain: 1

ARTICLE III, Section I
Section I  ++Both++ RCT and the events it both supports and organizes
will pursue ++both++ artistry and entertainment. RCTs goal

5) APPROVED. Aye: 3, Nay: 0, Abstain: 1

Article VI, Section VI
Section VI - If ^any member^:[++a general member++] has a complaint
regarding the professional conduct, fulfillment of duty, or competence
of ^any other member(s) (or, indeed, the entire board)^:[++an executive
member++], they may, and are encouraged to, issue a written complaint to
any member of the executive board.